+++ ios/Classes/handler/CommandHandler.swift
@@ -11,6 +11,8 @@ enum MethodName: String {
     case rotateVideo = "rotateVideo"
     case generateThumbnail = "generateThumbnail"
     case compressVideo = "compressVideo"
+    case cancelOperation = "cancelOperation"
+    case getVideoMetadata = "getVideoMetadata"
 }
 
 class CommandHandler {
@@ -26,20 +28,22 @@ class CommandHandler {
             .cropVideo: CropVideoCommand(),
             .rotateVideo: RotateVideoCommand(),
             .generateThumbnail: GenerateThumbnailCommand(),
-            .compressVideo: CompressVideoCommand()
+            .compressVideo: CompressVideoCommand(),
+            .cancelOperation: CancelOperationCommand(),
+            .getVideoMetadata: GetVideoMetadataCommand()
         ]
     }
     
     func handleCommand(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-        guard let methodName = MethodName(rawValue: call.method) else {
-            result(FlutterMethodNotImplemented)
-            return
-        }
-        
-        if let handler = handlers[methodName] {
-            handler.execute(call: call, result: result)
-        } else {
-            result(FlutterMethodNotImplemented)
+            guard let methodName = MethodName(rawValue: call.method) else {
+                result(FlutterMethodNotImplemented)
+                return
+            }
+            
+            if let handler = handlers[methodName] {
+                handler.execute(call: call, result: result)
+            } else {
+                result(FlutterMethodNotImplemented)
+            }
         }
     }
-}